+++ client/package.json
@@ -8,6 +8,8 @@
     "moment": "^2.24.0",
     "react": "^16.8.4",
     "react-big-calendar": "^0.20.3",
+    "react-dnd": "^7.3.2",
+    "react-dnd-html5-backend": "^7.2.0",
     "react-dom": "^16.8.4",
     "react-redux": "^6.0.1",
     "react-router-dom": "^4.3.1",
